+++ tests/MockAzureOpenAIEnv.py
@@ -2,8 +2,9 @@ from openai import AzureOpenAI
 
 
 class MockAzureOpenAI(AzureOpenAI):
-    def __init__(self):
-        pass
+    def __init__(self, json_error: bool = False, general_error: bool = False):
+        self.json_error = json_error
+        self.general_error = general_error
 
     @property
     def chat(self):
@@ -14,6 +15,10 @@ class MockAzureOpenAI(AzureOpenAI):
         return self
 
     def create(self, model, messages, temperature, response_format):
+        if self.json_error:
+            return MockAzureJSONError()
+        if self.general_error:
+            raise Exception("General error")
         return MockAzureOpenAIResponse()
 
 
@@ -33,6 +38,16 @@ class MockChoice:
 class MockAzureOpenAIResponse:
     def __init__(self):
         self.choices = [MockChoice('{"Categorie1": "Beloop1","Categorie2": "Beloop2"}')]
+        self.created = 1626161224
+        self.id = "cmpl-3LwvXrU5W6iY0Gd9ZJf9eY4v"
+        self.model = "aiva-gpt"
+
+
+class MockAzureJSONError:
+    """Alternative response, triggers JSONError"""
+
+    def __init__(self):
+        self.choices = [MockChoice('{"Categorie1": "Beloop1","Categorie2"= "Beloop2"}')]
         self.created = 1626161224
         self.id = "cmpl-3LwvXrU5W6iY0Gd9ZJf9eY4v"
         self.model = "aiva-gpt"